GenAI_Contract_Insights/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.37.0
pydantic==2.8.2
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
openai==1.35.14
